move code completion code from adapter to code compeltion parser
