+++ python/sglang/srt/code_completion_parser.py
@@ -15,12 +15,15 @@
 
 
 import dataclasses
+import json
 import logging
+import os
 from enum import auto
 
 from sglang.srt.openai_api.protocol import ChatCompletionRequest
 
 logger = logging.getLogger(__name__)
+completion_template_name = None
 
 
 class FimPosition:
@@ -54,6 +57,46 @@ class CompletionTemplate:
 completion_templates: dict[str, CompletionTemplate] = {}
 
 
+def load_completion_template_for_openai_api(tokenizer_manager, completion_template_arg):
+    global completion_template_name
+
+    logger.info(
+        f"Use completion template for the OpenAI-compatible API server: {completion_template_arg}"
+    )
+
+    if not completion_template_exists(completion_template_arg):
+        if not os.path.exists(completion_template_arg):
+            raise RuntimeError(
+                f"Completion template {completion_template_arg} is not a built-in template name "
+                "or a valid completion template file path."
+            )
+
+        assert completion_template_arg.endswith(
+            ".json"
+        ), "unrecognized format of completion template file"
+        with open(completion_template_arg, "r") as filep:
+            template = json.load(filep)
+            try:
+                fim_position = FimPosition[template["fim_position"]]
+            except KeyError:
+                raise ValueError(
+                    f"Unknown fim position: {template['fim_position']}"
+                ) from None
+            register_completion_template(
+                CompletionTemplate(
+                    name=template["name"],
+                    fim_begin_token=template["fim_begin_token"],
+                    fim_middle_token=template["fim_middle_token"],
+                    fim_end_token=template["fim_end_token"],
+                    fim_position=fim_position,
+                ),
+                override=True,
+            )
+        completion_template_name = template["name"]
+    else:
+        completion_template_name = completion_template_arg
+
+
 def register_completion_template(template: CompletionTemplate, override: bool = False):
     """Register a new completion template."""
     if not override:
@@ -68,15 +111,17 @@ def completion_template_exists(template_name: str) -> bool:
     return template_name in completion_templates
 
 
-def generate_completion_prompt_from_request(
-    request: ChatCompletionRequest, template_name: str
-) -> str:
-    return generate_completion_prompt(request.prompt, request.suffix, template_name)
+def generate_completion_prompt_from_request(request: ChatCompletionRequest) -> str:
+    global completion_template_name
+    if (completion_template_name == None) or (request.suffix == ""):
+        return request.prompt
+
+    return generate_completion_prompt(
+        request.prompt, request.suffix, completion_template_name
+    )
 
 
 def generate_completion_prompt(prompt: str, suffix: str, template_name: str) -> str:
-    if (suffix is None) or (suffix == ""):
-        return prompt
 
     completion_template = completion_templates[template_name]
     fim_begin_token = completion_template.fim_begin_token
